Renamed cloudresourceconetxt to could_resource_context
