+++ tests/integrations/cloud_resource_context/test_cloud_resource_context.py
@@ -4,7 +4,10 @@ import pytest
 import mock
 from mock import MagicMock
 
-from sentry_sdk.integrations.cloudresourcecontext import CLOUD_PLATFORM, CLOUD_PROVIDER
+from sentry_sdk.integrations.cloud_resource_context import (
+    CLOUD_PLATFORM,
+    CLOUD_PROVIDER,
+)
 
 AWS_EC2_EXAMPLE_IMDSv2_PAYLOAD = {
     "accountId": "298817902971",
@@ -85,7 +88,7 @@ GCP_GCE_EXAMPLE_METADATA_PLAYLOAD = {
 
 
 def test_is_aws_http_error():
-    from sentry_sdk.integrations.cloudresourcecontext import (
+    from sentry_sdk.integrations.cloud_resource_context import (
         CloudResourceContextIntegration,
     )
 
@@ -100,7 +103,7 @@ def test_is_aws_http_error():
 
 
 def test_is_aws_ok():
-    from sentry_sdk.integrations.cloudresourcecontext import (
+    from sentry_sdk.integrations.cloud_resource_context import (
         CloudResourceContextIntegration,
     )
 
@@ -120,7 +123,7 @@ def test_is_aws_ok():
 
 
 def test_is_aw_exception():
-    from sentry_sdk.integrations.cloudresourcecontext import (
+    from sentry_sdk.integrations.cloud_resource_context import (
         CloudResourceContextIntegration,
     )
 
@@ -167,7 +170,7 @@ def test_is_aw_exception():
     ],
 )
 def test_get_aws_context(http_status, response_data, expected_context):
-    from sentry_sdk.integrations.cloudresourcecontext import (
+    from sentry_sdk.integrations.cloud_resource_context import (
         CloudResourceContextIntegration,
     )
 
@@ -182,7 +185,7 @@ def test_get_aws_context(http_status, response_data, expected_context):
 
 
 def test_is_gcp_http_error():
-    from sentry_sdk.integrations.cloudresourcecontext import (
+    from sentry_sdk.integrations.cloud_resource_context import (
         CloudResourceContextIntegration,
     )
 
@@ -197,7 +200,7 @@ def test_is_gcp_http_error():
 
 
 def test_is_gcp_ok():
-    from sentry_sdk.integrations.cloudresourcecontext import (
+    from sentry_sdk.integrations.cloud_resource_context import (
         CloudResourceContextIntegration,
     )
 
@@ -212,7 +215,7 @@ def test_is_gcp_ok():
 
 
 def test_is_gcp_exception():
-    from sentry_sdk.integrations.cloudresourcecontext import (
+    from sentry_sdk.integrations.cloud_resource_context import (
         CloudResourceContextIntegration,
     )
 
@@ -256,7 +259,7 @@ def test_is_gcp_exception():
     ],
 )
 def test_get_gcp_context(http_status, response_data, expected_context):
-    from sentry_sdk.integrations.cloudresourcecontext import (
+    from sentry_sdk.integrations.cloud_resource_context import (
         CloudResourceContextIntegration,
     )
 
@@ -282,7 +285,7 @@ def test_get_gcp_context(http_status, response_data, expected_context):
     ],
 )
 def test_get_cloud_provider(is_aws, is_gcp, expected_provider):
-    from sentry_sdk.integrations.cloudresourcecontext import (
+    from sentry_sdk.integrations.cloud_resource_context import (
         CloudResourceContextIntegration,
     )
 
@@ -302,7 +305,7 @@ def test_get_cloud_provider(is_aws, is_gcp, expected_provider):
     ],
 )
 def test_get_cloud_resource_context_unsupported_providers(cloud_provider):
-    from sentry_sdk.integrations.cloudresourcecontext import (
+    from sentry_sdk.integrations.cloud_resource_context import (
         CloudResourceContextIntegration,
     )
 
@@ -321,7 +324,7 @@ def test_get_cloud_resource_context_unsupported_providers(cloud_provider):
     ],
 )
 def test_get_cloud_resource_context_supported_providers(cloud_provider):
-    from sentry_sdk.integrations.cloudresourcecontext import (
+    from sentry_sdk.integrations.cloud_resource_context import (
         CloudResourceContextIntegration,
     )
 
@@ -350,7 +353,7 @@ def test_get_cloud_resource_context_supported_providers(cloud_provider):
 def test_setup_once(
     cloud_provider, cloud_resource_context, warning_called, set_context_called
 ):
-    from sentry_sdk.integrations.cloudresourcecontext import (
+    from sentry_sdk.integrations.cloud_resource_context import (
         CloudResourceContextIntegration,
     )
 
@@ -360,10 +363,10 @@ def test_setup_once(
     )
 
     with mock.patch(
-        "sentry_sdk.integrations.cloudresourcecontext.set_context"
+        "sentry_sdk.integrations.cloud_resource_context.set_context"
     ) as fake_set_context:
         with mock.patch(
-            "sentry_sdk.integrations.cloudresourcecontext.logger.warning"
+            "sentry_sdk.integrations.cloud_resource_context.logger.warning"
         ) as fake_warning:
             CloudResourceContextIntegration.setup_once()
