+++ tox.ini
@@ -50,7 +50,7 @@ envlist =
     {py3.6,py3.7,py3.8}-chalice-v{1.16,1.17,1.18,1.19,1.20}
 
     # Cloud Resource Context
-    {py3.6,py3.7,py3.8,py3.9,py3.10,py3.11}-cloudresourcecontext
+    {py3.6,py3.7,py3.8,py3.9,py3.10,py3.11}-cloudre_source_context
 
     # Django
     # - Django 1.x
@@ -386,7 +386,7 @@ setenv =
     bottle: TESTPATH=tests/integrations/bottle
     celery: TESTPATH=tests/integrations/celery
     chalice: TESTPATH=tests/integrations/chalice
-    cloudresourcecontext: TESTPATH=tests/integrations/cloudresourcecontext
+    cloud_resource_context: TESTPATH=tests/integrations/cloud_resource_context
     django: TESTPATH=tests/integrations/django
     falcon: TESTPATH=tests/integrations/falcon
     fastapi:  TESTPATH=tests/integrations/fastapi
